+++ priam/src/test/java/com/netflix/priam/backup/FakedS3EncryptedFileSystem.java
@@ -98,7 +98,7 @@ public class FakedS3EncryptedFileSystem implements IBackupFileSystem {
 	}
 
 	@Override
-	public int getAWSSlowDownExceptionCounter() {
+	public long getAWSSlowDownExceptionCounter() {
 		return 0;
 	}
